refactor: 컴포넌트 분리 및 컴포넌트 Index.ts 설정
