--- src/shared/ui/card/Card.tsx
@@ -1,15 +1,0 @@
-import { forwardRef } from 'react';
-
-interface CardProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-
-export const Card = forwardRef<HTMLDivElement, CardProps>(({ className, ...props }, ref) => (
-  <div
-    ref={ref}
-    className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`}
-    {...props}
-  />
-));
-
-Card.displayName = 'Card';
