--- src/shared/ui/card/CardContent.tsx
@@ -1,13 +1,0 @@
-import { forwardRef } from 'react';
-
-interface CardContentProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-
-export const CardContent = forwardRef<HTMLDivElement, CardContentProps>(
-  ({ className, ...props }, ref) => (
-    <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
-  ),
-);
-
-CardContent.displayName = 'CardContent';
