--- src/shared/ui/card/CardHeader.tsx
@@ -1,13 +1,0 @@
-import { forwardRef } from 'react';
-
-interface CardHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
-  className?: string;
-}
-
-export const CardHeader = forwardRef<HTMLDivElement, CardHeaderProps>(
-  ({ className, ...props }, ref) => (
-    <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
-  ),
-);
-
-CardHeader.displayName = 'CardHeader';
