+++ src/shared/ui/index.tsx
@@ -1,0 +1,7 @@
+export { Button } from './Button';
+export { Input } from './Input';
+export { Dialog } from './Dialog';
+export { Select } from './Select';
+export { Table } from './Table';
+export { Textarea } from './Textarea';
+export { Card } from './Card';
